Setup endpoints and exceptions

+++ src/main/java/com/example/RecipeBookBackend/Service/RecipeBookService.java
@@ -1,11 +1,17 @@
 package com.example.RecipeBookBackend.Service;
 
+import com.example.RecipeBookBackend.Exceptions.RecipeNotFoundException;
+import com.example.RecipeBookBackend.Model.Ingredients;
 import com.example.RecipeBookBackend.Model.Recipe;
+import com.example.RecipeBookBackend.Model.Steps;
+import com.example.RecipeBookBackend.Repository.IngredientsRepository;
 import com.example.RecipeBookBackend.Repository.RecipeBookRepository;
+import com.example.RecipeBookBackend.Repository.StepsRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,9 +22,27 @@ public class RecipeBookService {
     @Autowired
     RecipeBookRepository recipeBookRepository;
 
+    @Autowired
+    StepsRepository stepsRepository;
+
+    @Autowired
+    IngredientsRepository ingredientsRepository;
+
     //CREATE
     public void addRecipe(Recipe recipe) {
         recipeBookRepository.save(recipe);
+    }
+
+    public void addSteps(ArrayList<Steps> steps) {
+        Recipe correctRecipeId = getLastRecipeID();
+        steps.forEach((step -> step.setRecipe(correctRecipeId)));
+        stepsRepository.saveAll(steps);
+    }
+
+    public void addIngredients(ArrayList<Ingredients> ingredients) {
+        Recipe correctRecipeId = getLastRecipeID();
+        ingredients.forEach((step -> step.setRecipe(correctRecipeId)));
+        ingredientsRepository.saveAll(ingredients);
     }
 
     //READ
@@ -34,16 +58,28 @@ public class RecipeBookService {
         Optional<Recipe> recipe = recipeBookRepository.findById(id);
 
         if(recipe.isEmpty()) {
-            throw new Error();
+            throw new RecipeNotFoundException();
         }
 
         return recipe.get();
     }
 
+    public Recipe getLastRecipeID() {
+        return recipeBookRepository.getLastRecipeID();
+    }
+
+    public List<String> getIngredientsByID(long id) {
+        return recipeBookRepository.getIngredientsByID(id);
+    }
+
+    public List<String> getStepsByID(long id) {
+        return recipeBookRepository.getStepsByID(id);
+    }
+
     //UPDATE
     public void updateRecipe(Recipe newRecipe, long id) {
         if(!recipeBookRepository.existsById(id)) {
-            throw new Error();
+            throw new RecipeNotFoundException();
         }
 
         newRecipe.setId(id);
@@ -54,6 +90,10 @@ public class RecipeBookService {
     //DELETE
     @Transactional
     public void deleteRecipe(long id) {
+        if(!recipeBookRepository.existsById(id)) {
+            throw new RecipeNotFoundException();
+        }
+
         recipeBookRepository.deleteRecipeByid(id);
     }
 }
